Separated lekvar into separate files
